Support system and Properties parsing from JS
