+++ core/jvm/src/main/scala/japgolly/clearconfig/internals/StoreJvm.scala
@@ -1,34 +1,5 @@
 package japgolly.clearconfig.internals
 
-import cats.Applicative
-import java.io.InputStream
-import java.util.Properties
-import scala.jdk.CollectionConverters._
-
 object StoreJvm extends StoreObjectJvm
 trait StoreObjectJvm extends StoreObject {
-
-  private def propsToMap(p: Properties): Map[Key, String] =
-    p.keys().asScala.map { kx =>
-      val k = "" + kx
-      Key(k) -> p.getProperty(k)
-    }.toMap
-
-  final def ofJavaProps[F[_]](p: Properties)(implicit F: Applicative[F]): Store[F] = {
-    lazy val m = propsToMap(p)
-    apply(F.point(m))
-  }
-
-  final def ofJavaPropsFromInputStream[F[_]](is: InputStream, close: Boolean = true)(implicit F: Applicative[F]): Store[F] = {
-    lazy val m =
-      try {
-        val p = new Properties()
-        p.load(is)
-        propsToMap(p)
-      } finally
-        if (close)
-          is.close()
-    apply(F.point(m))
-  }
-
 }
