+++ core/shared/src/main/scala/japgolly/clearconfig/internals/SourceName.scala
@@ -10,6 +10,7 @@ trait SourceNameObject {
   final def apply(name: String) = new SourceName(name)
   final def default = SourceName("Default")
   final def environment = SourceName("Env")
+  final def system = SourceName("System")
 
 }
